// src/app/modules/authentication/helpers/enums/fnz-actividad.enum.ts
export enum FnzEActividad {
    FIN = 0,
    NUEVO_FOLIO = 1,
    SOLICITUD = 2,
    CARGA_DOCUMENTAL = 3,
    VALIDACION_DIGITAL = 4,
    GENERACION_FORMATOS = 28,
    FIRMA_DOCUMENTAL = 5,
    VALIDACION_FIRMAS = 7,
    VALIDACION_AFIANZADORA = 29,
    RECOLECCION_DE_FISICOS = 18,
    VALIDACION_DE_ORIGINALES = 19,
    CONFIRMACION_ENTREGA = 8,
    CONTACTO_TELEFONICO = 9,
    CONTACTO_ASEGURADORA = 10,
}