// src/app/shared/helpers/interfaces/response-file.ts
export interface responseFile {
    aseguradora:    string;
    titular:        string;
    documento:      string;
    clave:      string;
    nombreOriginal: string;
    nombreCorto:    string;
    version:        number;
    url:            string;
    usuarioAlta:    string;
    eliminado:      boolean;
    _id:            string;
    fechaHoraAlta:  Date;
}